Finish work on the interface
