+++ Lara/Theme/Sizes.swift
@@ -10,6 +10,8 @@ import SwiftUI
 struct Sizes {
     static let elementWidth: CGFloat = 327
     static let nearDoctorWidth: CGFloat = 295
+    static let cornerRadius: CGFloat = 12
+    static let tabItemMinWidth: CGFloat = 70
 }
 
 extension View {
